unit tests from regex_manager.rs moved to tests/unit/regex_manager.rs
